+++ tests/file/UploadTest.php
@@ -77,4 +77,41 @@ class UploadTest extends TestCase
         $this->assertEquals(0, $collection->count());
         $this->assertEquals(0, $collection->getChunkCollection()->count());
     }
+
+    /**
+     * @see https://github.com/yiisoft/yii2-mongodb/issues/179
+     *
+     * @depends testAddContent
+     */
+    public function testCustomId()
+    {
+        $collection = $this->getConnection()->getFileCollection();
+
+        $id = new ObjectID();
+        $upload = $collection->createUpload([
+            'document' => [
+                '_id' => $id,
+            ]
+        ]);
+        $document = $upload->addContent('object ID')->complete();
+        $this->assertSame($id, $document['_id']);
+
+        $id = '5889da846748383ce0556bd9';
+        $upload = $collection->createUpload([
+            'document' => [
+                '_id' => $id,
+            ]
+        ]);
+        $document = $upload->addContent('object ID')->complete();
+        $this->assertSame($id, (string)$document['_id']);
+
+        $id = 'custom-id';
+        $upload = $collection->createUpload([
+            'document' => [
+                '_id' => $id,
+            ]
+        ]);
+        $document = $upload->addContent('object ID')->complete();
+        $this->assertSame($id, $document['_id']);
+    }
 }